refactor: removed newly introduced duplications
