+++ rust/sealevel/libraries/test-utils/Cargo.toml
@@ -16,6 +16,7 @@ spl-noop.workspace = true
 spl-token-2022.workspace = true
 
 hyperlane-core = { path = "../../../main/hyperlane-core" }
+k256.workspace = true
 hyperlane-sealevel-interchain-security-module-interface = { path = "../interchain-security-module-interface" }
 hyperlane-sealevel-mailbox = { path = "../../programs/mailbox", features = [
     "no-entrypoint",
